@@ -11,18 +11,19 @@
     "@testing-library/react": "11.2.5",
     "@testing-library/user-event": "12.6.2",
     "chalk": "4.1.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1"
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
+    "eslint": "7.32.0",
     "eslint-config-airbnb": "18.2.1",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-prettier": "3.3.1",
     "husky": "4.3.8",
-    "lint-staged": "10.5.4",
+    "lint-staged": "6.1.1",
     "npm-run-all": "4.1.5",
     "prettier": "2.2.1",
-    "react-scripts": "4.0.3"
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
